fix josh comments
