+++ backend/internal/storage/postgres/schema/child/update_by_id.go
@@ -3,17 +3,12 @@ package child
 import (
 	"context"
 	"errors"
-	"net/http"
 	"skillspark/internal/errs"
 	"skillspark/internal/models"
 	"skillspark/internal/storage/postgres/schema"
-	"skillspark/internal/storage/postgres/testutil"
-	"skillspark/internal/utils"
-	"testing"
 
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
-	"github.com/stretchr/testify/assert"
 )
 
 func (r *ChildRepository) UpdateChildByID(ctx context.Context, childID uuid.UUID, child *models.UpdateChildInput) (*models.Child, error) {
@@ -59,31 +54,4 @@ func (r *ChildRepository) UpdateChildByID(ctx context.Context, childID uuid.UUID
 	}
 
 	return &updatedChild, nil
-}
-
-func TestChildRepository_UpdateChildByID_NotFound(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping database test in short mode")
-	}
-
-	testDB := testutil.SetupTestDB(t)
-	repo := NewChildRepository(testDB)
-	ctx := context.Background()
-
-	nonExistentID := uuid.MustParse("00000000-0000-0000-0000-000000000000")
-
-	input := &models.UpdateChildInput{}
-	input.Body.Name = utils.PtrString("Updated Name")
-	input.Body.BirthMonth = utils.PtrInt(5)
-	input.Body.BirthYear = utils.PtrInt(2019)
-
-	child, err := repo.UpdateChildByID(ctx, nonExistentID, input)
-
-	assert.Nil(t, child)
-	assert.NotNil(t, err)
-
-	httpErr, ok := err.(*errs.HTTPError)
-	assert.True(t, ok, "expected *errs.HTTPError")
-
-	assert.Equal(t, http.StatusNotFound, httpErr.Code)
 }
