+++ backend/internal/storage/postgres/schema/child/update_by_id_test.go
@@ -2,11 +2,15 @@ package child
 
 import (
 	"context"
+	"net/http"
 	"testing"
 
+	"skillspark/internal/errs"
 	"skillspark/internal/models"
 	"skillspark/internal/storage/postgres/testutil"
+	"skillspark/internal/utils"
 
+	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -49,4 +53,31 @@ func TestChildRepository_UpdateChildByID(t *testing.T) {
 
 	assert.False(t, updatedChild.UpdatedAt.IsZero())
 	assert.False(t, updatedChild.CreatedAt.IsZero())
+}
+
+func TestChildRepository_UpdateChildByID_NotFound(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping database test in short mode")
+	}
+
+	testDB := testutil.SetupTestDB(t)
+	repo := NewChildRepository(testDB)
+	ctx := context.Background()
+
+	nonExistentID := uuid.MustParse("00000000-0000-0000-0000-000000000000")
+
+	input := &models.UpdateChildInput{}
+	input.Body.Name = utils.PtrString("Updated Name")
+	input.Body.BirthMonth = utils.PtrInt(5)
+	input.Body.BirthYear = utils.PtrInt(2019)
+
+	child, err := repo.UpdateChildByID(ctx, nonExistentID, input)
+
+	assert.Nil(t, child)
+	assert.NotNil(t, err)
+
+	httpErr, ok := err.(*errs.HTTPError)
+	assert.True(t, ok, "expected *errs.HTTPError")
+
+	assert.Equal(t, http.StatusNotFound, httpErr.Code)
 }
